src/nnhub/requirements.txt
aiosqlite==0.19.0 ; python_version >= "3.9" and python_version < "4.0"
annotated-types==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.9" and python_version < "4.0"
certifi==2024.2.2 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
exceptiongroup==1.2.0 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.104.1 ; python_version >= "3.9" and python_version < "4.0"
filelock==3.13.1 ; python_version >= "3.9" and python_version < "4.0"
fsspec==2024.2.0 ; python_version >= "3.9" and python_version < "4.0"
gunicorn==21.2.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
huggingface-hub==0.21.3 ; python_version >= "3.9" and python_version < "4.0"
idna==3.6 ; python_version >= "3.9" and python_version < "4.0"
lorem==0.1.1 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.26.3 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.9" and python_version < "4.0"
pillow==10.2.0 ; python_version >= "3.9" and python_version < "4.0"
pydantic-core==2.14.5 ; python_version >= "3.9" and python_version < "4.0"
pydantic-settings==2.1.0 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.5.2 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
regex==2023.12.25 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
safetensors==0.4.2 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.9" and python_version < "4.0"
tokenizers==0.15.2 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.66.2 ; python_version >= "3.9" and python_version < "4.0"
transformers==4.38.2 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
uvicorn==0.24.0.post1 ; python_version >= "3.9" and python_version < "4.0"
aiofiles
aiohttp
aioredis==1.3.1